feat: add option to declare to skip already declared class
Before this commit, if a class was already declared, an error was raised.
But sometimes, the user may want to continue even if the class is
already declared.
This commit introduces a new option skip_if_declared, which add
a verification of the class declaration and skip if the class
is already declared without sending an error.